Refactor event handling from UserRegisteredEvent to UserRegistered
Updated AuthController.cs to publish UserRegistered instead of UserRegisteredEvent. Removed UserRegisteredEvent.cs. Modified Program.cs to update MassTransit configuration for UserRegistered. Updated HolaViaje.Social.csproj to change file paths and add a new folder for IntegrationEvents\Consumers. Changed IdentityEventsConsumer.cs to use UserRegistered and moved it to a new namespace. Added new UserRegistered.cs file defining the UserRegistered class.
