+++ src/HolaViaje.Account/Features/Identity/Events/UserRegistered.cs
@@ -1,6 +1,6 @@
 ﻿namespace HolaViaje.Account.Features.Identity.Events;
 
-public class UserRegisteredEvent
+public class UserRegistered
 {
     public long AccountId { get; set; }
     public string Email { get; set; } = string.Empty;
